throttling support implemented

+++ Src/Library/Endpoint/Endpoint.Setup.cs
@@ -305,7 +305,7 @@ public abstract partial class Endpoint<TRequest, TResponse> : BaseEndpoint where
     /// <param name="hitLimit">how many requests are allowed within the given duration</param>
     /// <param name="durationSeconds">the frequency in seconds where the accrued hit count should be reset</param>
     /// <param name="headerName">the name of the request header used to uniquely identify clients</param>
-    protected void Throttle(int hitLimit, double durationSeconds, string headerName = "X-Forwarded-For")
+    protected void Throttle(int hitLimit, int durationSeconds, string headerName = "X-Forwarded-For")
     {
         Settings.HitCounter = new(headerName, durationSeconds, hitLimit);
     }
